+++ packages/components/organisms/f-checkout/src/components/Selector.vue
@@ -44,7 +44,8 @@ export default {
     computed: {
         ...mapState(VUEX_CHECKOUT_MODULE, [
             'availableFulfilment',
-            'serviceType'
+            'serviceType',
+            'time'
         ]),
 
         orderMethod () {
@@ -92,7 +93,7 @@ export default {
     },
 
     mounted () {
-        this.initFulfilmentTime(this.fulfilmentTimes);
+        this.initFulfilmentTime();
     },
 
     methods: {
@@ -124,20 +125,24 @@ export default {
         },
 
         /**
-         * Initialise the first set of times so if the user proceeds with the preset delivery time
-         * we have access to the from/to times up front.
-         *
-         * @param times
+         * If we have previously set a time then initialise the `selected dropdown item` with that value; if
+         * currently present in the Fulfilment times array, otherwise initialise the with first value in the
+         * Fulfilment times array.
+         * Note; we always need to pre-select a time; if available, so the user can always proceeds with a selected time.
          */
-        initFulfilmentTime (times) {
-            if (times.length && times[0].value) {
-                this.updateFulfilmentTime({
-                    from: times[0].value,
-                    to: times[0].value
-                });
+        initFulfilmentTime () {
+            const times = this.fulfilmentTimes || [];
 
-                this.selectedAvailableFulfilmentTime = times[0].value;
-            }
+            let selectedTime = (times.some(i => i.value === this.time.from) && this.time.from)
+                                || (times.length && times[0].value);
+            selectedTime = selectedTime === 0 ? '' : selectedTime;
+
+            this.updateFulfilmentTime({
+                from: selectedTime,
+                to: selectedTime
+            });
+
+            this.selectedAvailableFulfilmentTime = selectedTime;
         },
 
         /**